Перенесел функцию в отдельный файл helpers
