+++ requirements.txt
@@ -5,3 +5,4 @@ passlib==1.7.4
 pyjwt==2.9.0
 asyncpg==0.29.0
 uvicorn==0.31.0
+pydantic[email]
