+++ src/config/configs.py
@@ -1,0 +1,5 @@
+import os
+
+SECRET_KEY = os.getenv("SECRET_KEY", "your_default_secret_key")
+ALGORITHM = os.getenv("ALGORITHM", "HS256")
+ACCESS_TOKEN_EXPIRE_MINUTES = int(os.getenv("ACCESS_TOKEN_EXPIRE_MINUTES", 30))
